Fix codereview comments and warnings.
