+++ source/slang/slang-ir-util.cpp
@@ -1,0 +1,29 @@
+#include "slang-ir-util.h"
+
+namespace Slang
+{
+
+bool isPointerOfType(IRInst* type, IROp opCode)
+{
+    if (auto ptrType = as<IRPtrTypeBase>(type))
+    {
+        return ptrType->getValueType() && ptrType->getValueType()->getOp() == opCode;
+    }
+    return false;
+}
+
+bool isPointerOfType(IRInst* type, IRInst* elementType)
+{
+    if (auto ptrType = as<IRPtrTypeBase>(type))
+    {
+        return ptrType->getValueType() && isTypeEqual(ptrType->getValueType(), (IRType*)elementType);
+    }
+    return false;
+}
+
+bool isPtrToClassType(IRInst* type)
+{
+    return isPointerOfType(type, kIROp_ClassType);
+}
+
+}
