+++ source/slang/slang-ir-util.h
@@ -1,0 +1,22 @@
+// slang-ir-util.h
+#ifndef SLANG_IR_UTIL_H_INCLUDED
+#define SLANG_IR_UTIL_H_INCLUDED
+
+// This file contains utility functions for operating with Slang IR.
+//
+#include "slang-ir.h"
+
+namespace Slang
+{
+
+bool isPtrToClassType(IRInst* type);
+
+// True if ptrType is a pointer type to elementType
+bool isPointerOfType(IRInst* ptrType, IRInst* elementType);
+
+// True if ptrType is a pointer type to a type of opCode
+bool isPointerOfType(IRInst* ptrType, IROp opCode);
+
+}
+
+#endif
